requirements.txt
langchain
langchain-community
protobuf
transformers>=4.37.2
cpm_kernels
torch>=2.0
gradio>=3.38.0,<4.0.0
mdtex2html
sentencepiece
accelerate>=0.27.2
pydantic[email]
fastapi
python-dotenv
uvicorn[standard]
sentence-transformers
python-multipart
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy
datasets>=2.14.3
peft>=0.10.0
trl>=0.8.1
scipy
einops
sse-starlette
matplotlib
fire
galore-torch
loguru
tiktoken
chromadb